+++ monorepo/gh-pages/gh-pages-config-dev-docs/config/markdown.php
@@ -71,11 +71,12 @@ return [
     |
     | This feature allows you to use basic Laravel Blade in Markdown files.
     |
-    | It's disabled by default since can be a security risk as it allows
-    | arbitrary PHP to run. But if your Markdown is trusted, try it out!
+    | BladeDown is enabled by default because source files in Hyde projects are
+    | generally trusted and reviewed. Since Blade can execute arbitrary PHP,
+    | disable this when compiling untrusted or unreviewed Markdown.
     |
     | To see the syntax and usage, see the documentation:
-    | @see https://hydephp.com/docs/master/advanced-markdown#blade-support
+    | @see https://hydephp.com/docs/3.x/advanced-markdown#using-blade-in-markdown
     |
     */